requirements.txt
python-dotenv
ebooklib
beautifulsoup4
lxml
pandas
numpy
langchain
langchain_community
langchain-openai
langchain-groq
matplotlib
scipy
networkx
openai
diskcache
pydub
streamlit
python-slugify